fix: 산책 중 거리 계산 km로 수정(#118)
